requirements.txt
flask==2.3.3
python-docx==0.8.11
pillow==10.0.0
qrcode==7.4.2
numpy==1.24.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
colorama==0.4.6